Research/requirements.txt
crossrefapi